+++ src/components/Heroes/rankColumns.ts
